requirements.txt
pytest==7.4.3
selene>=2.0.0rc4
requests==2.31.0
jsonschema==4.20.0
allure-pytest==2.13.2
Appium-Python-Client==3.1.1
python-dotenv==1.0.0
pydantic==2.5.2